+++ src/MyMediaLite/Data/TimedRatingsProxy.cs
@@ -1,4 +1,4 @@
-// Copyright (C) 2011 Zeno Gantner
+// Copyright (C) 2011, 2012 Zeno Gantner
 //
 // This file is part of MyMediaLite.
 //
@@ -33,7 +33,7 @@ namespace MyMediaLite.Data
 		{
 			Users  = new ListProxy<int>(ratings.Users, indices);
 			Items  = new ListProxy<int>(ratings.Items, indices);
-			Values = new ListProxy<double>(ratings, indices);
+			Values = new ListProxy<float>(ratings, indices);
 			Times  = new ListProxy<DateTime>(ratings.Times, indices);
 
 			MaxUserID = ratings.MaxUserID;
