+++ src/MyMediaLite/RatingPrediction/UserKNN.cs
@@ -1,4 +1,4 @@
-// Copyright (C) 2010, 2011 Zeno Gantner
+// Copyright (C) 2010, 2011, 2012 Zeno Gantner
 //
 // This file is part of MyMediaLite.
 //
@@ -91,7 +91,7 @@ namespace MyMediaLite.RatingPrediction
 		abstract protected void RetrainUser(int user_id);
 
 		///
-		public override void AddRating(int user_id, int item_id, double rating)
+		public override void AddRating(int user_id, int item_id, float rating)
 		{
 			baseline_predictor.AddRating(user_id, item_id, rating);
 			data_user[user_id, item_id] = true;
@@ -99,7 +99,7 @@ namespace MyMediaLite.RatingPrediction
 		}
 
 		///
-		public override void UpdateRating(int user_id, int item_id, double rating)
+		public override void UpdateRating(int user_id, int item_id, float rating)
 		{
 			baseline_predictor.UpdateRating(user_id, item_id, rating);
 			RetrainUser(user_id);
